Make sure the metadata is preserved in the correct way
Until now we tried to read the metadata from the API call,
but it seems that its not honored when we load the squashed
image into Docker, some properities (like docker_version) are
not applied. Right now we use the metadata from exported image.

BTW, Docker, why oh why we have mixed underscores with CamleCase
in the JSON file?!

Fixes #28.