Better handling of symlinks pointing to non-existing locations
If we have a symlink to a location that does not exist in the
squashed tar (yet), we don't know if it's a file or a directory.
In case it's a directory - we cannot add any new files under
this directory because extration will fail at the import time.
Assume here the target is directory (if it's a file it won't
break) and skip all files that should be added under this directory.

Fixes #104.