Make it possible to remove source image
When we run squashing sometimes it's desired to remove the original image.
This is especially important if we tag squashed image with the previous image
tag. This leads to "orphaned" layers that should be removed.

This commit adds a new --cleanup CLI switch to remove the original image. It's
disabled by default.

Fixes #61