Add back "broken" symbolic links if these were already broken
Currently we try to fix the issue when there is a symbolink link
pointing to a target that was removed - we remove the link too.
This can be an issue if someone intentionally creates a broken
symbolic link to be populated later for example by using volumes.

Fixes #112

Additionally all paths everywhere are normalized to make the comparison
actually work.