Marker file prevents symbolic link from being added back
When we have a directory structure and remove part of it in one
layer and add a symlink to removed directory in other layer --
this symlink is skipped. For example:

FROM busybox:1.24
RUN mkdir /dir
RUN touch /dir/a
RUN touch /dir/b
RUN mkdir /dir/dir
RUN touch /dir/dir/file
RUN mv /dir/dir /newdir
RUN ln -s /newdir /dir/dir

When we squash two latest layers then the dir/.wh.dir marker file is
found which prevents the symlinks from being added to the squashed
image because the target of this symlink is the same as the marker
files which suggests that it should be skipped.

Fixes #116