Reduce marker files to be added back to required minimum
In some cases, one marker file will overlap
with others making others not necessary.

This is not only about adding less marker files, but
if we try to add a marker file for a file or directory
deeper in the hierarchy of already marked directory,
the image will not be successfully loaded back into Docker
daemon.

Fixes #181