Handle opaque directories
Opaque directories are special case whiteout files. In case
an opaque directory is found in a layer, all content inside
of this rectory in this layer should be copied to the squashed
layer, but any content in layers below that layer on the
opaque directory path must be ignored.

Fixes #186